Fixed up ZilchShaders for 64bit.

+++ Libraries/ZilchShaders/ArithmeticInstructions.cpp
@@ -207,9 +207,9 @@ void RegisterArithmeticOps(ZilchSpirVFrontEnd* translator, ZilchShaderIRLibrary*
   }
 
   // Register all real matrix instructions.
-  for(size_t y = 2; y <= 4; ++y)
+  for(u32 y = 2; y <= 4; ++y)
   {
-    for(size_t x = 2; x <= 4; ++x)
+    for(u32 x = 2; x <= 4; ++x)
     {
       Zilch::BoundType* zilchType = types.GetMatrixType(y, x)->mZilchType;
       String zilchTypeName = zilchType->ToString();
@@ -222,7 +222,7 @@ void RegisterArithmeticOps(ZilchSpirVFrontEnd* translator, ZilchShaderIRLibrary*
       
       // Iterate over all of the other matrix dimensions to make the multiplication functions 
       // (e.g. Real2x3 * real3x2, Real2x3 * Real3x3, etc...)
-      for(size_t z = 2; z <= 4; ++z)
+      for(u32 z = 2; z <= 4; ++z)
       {
         Zilch::BoundType* rhsMatrixType = types.GetMatrixType(x, z)->mZilchType;
         mathTypeResolver.RegisterFunctionResolver(GetStaticFunction(mathType, "Multiply", zilchTypeName, rhsMatrixType->ToString()), ResolveMatrixTimesMatrix);
