+++ Libraries/ZilchShaders/SimpleZilchShaderIRGenerator.hpp
@@ -59,9 +59,9 @@ public:
   /// multi-to-multi mapping as any stage can split or merge some of these types together.
   struct SampledImageRemappingData
   {
-    Array<int> mSampledImageIds;
-    Array<int> mImageIds;
-    Array<int> mSamplerIds;
+    Array<size_t> mSampledImageIds;
+    Array<size_t> mImageIds;
+    Array<size_t> mSamplerIds;
   };
 
   /// Describes how to find the reflection data for a ssbo.
@@ -119,7 +119,7 @@ public:
   ShaderStageInterfaceReflection mReflection;
 
 private:
-  typedef HashMap<String, int> NameToIndexMap;
+  typedef HashMap<String, size_t> NameToIndexMap;
 
   // Internal helper class used to remap uniform buffers between stages to each other.
   struct SimpleResourceRemappingData
