+++ Libraries/ZilchShaders/ZilchSpirVDisassemblerBackend.cpp
@@ -26,7 +26,7 @@ bool ZilchSpirVDisassemblerBackend::RunTranslationPass(ShaderTranslationPassResu
 
   ShaderByteStream& inputByteStream = inputData.mByteStream;
   uint32_t* data = (uint32_t*)inputByteStream.Data();
-  uint32 wordCount = inputByteStream.WordCount();
+  size_t wordCount = inputByteStream.WordCount();
 
   spv_text text;
   spv_diagnostic diagnostic = nullptr;
